requirements.txt
backports.shutil-get-terminal-size==1.0.0
click==6.6
colorama==0.3.7
decorator==4.0.9
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
prettytable==0.7.2
ptyprocess==0.5.1
python-firebase==1.2
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
